requirements.txt
telethon==1.37.0
supabase==2.13.0
python-dotenv==1.1.0
aiohttp==3.11.18
fastapi==0.110.0
uvicorn==0.27.1
